+++ GitTfs/Commands/InitBranch.cs
@@ -188,7 +188,7 @@ namespace Sep.Git.Tfs.Commands
         {
             public string TfsRepositoryPath { get; set; }
             public string GitBranchNameExpected { get; set; }
-            public long RootChangesetId { get; set; }
+            public int RootChangesetId { get; set; }
         }
 
         [DebuggerDisplay("{TfsRepositoryPath} C{RootChangesetId}")]
@@ -197,7 +197,7 @@ namespace Sep.Git.Tfs.Commands
             public string TfsRepositoryPath { get; set; }
             public IGitTfsRemote TfsRemote { get; set; }
             public bool IsEntirelyFetched { get; set; }
-            public long RootChangesetId { get; set; }
+            public int RootChangesetId { get; set; }
             public IList<RootBranch> CreationBranchData { get; set; }
             public Exception Error { get; set; }
         }
