+++ GitTfs/Core/IGitRepository.cs
@@ -24,7 +24,7 @@ namespace Sep.Git.Tfs.Core
         void MoveTfsRefForwardIfNeeded(IGitTfsRemote remote);
         void MoveTfsRefForwardIfNeeded(IGitTfsRemote remote, string @ref);
         IEnumerable<TfsChangesetInfo> GetLastParentTfsCommits(string head);
-        TfsChangesetInfo GetTfsChangesetById(string remoteRef, long changesetId);
+        TfsChangesetInfo GetTfsChangesetById(string remoteRef, int changesetId);
         TfsChangesetInfo GetTfsCommit(GitCommit commit);
         TfsChangesetInfo GetTfsCommit(string sha);
         TfsChangesetInfo GetCurrentTfsCommit();
@@ -42,7 +42,7 @@ namespace Sep.Git.Tfs.Core
         string AssertValidBranchName(string gitBranchName);
         bool CreateBranch(string gitBranchName, string target);
         Branch RenameBranch(string oldName, string newName);
-        string FindCommitHashByChangesetId(long changesetId);
+        string FindCommitHashByChangesetId(int changesetId);
         void CreateTag(string name, string sha, string comment, string Owner, string emailOwner, System.DateTime creationDate);
         void CreateNote(string sha, string content, string owner, string emailOwner, DateTime creationDate);
         void MoveRemote(string oldRemoteName, string newRemoteName);
